requirements.txt
blinker==1.7.0
boto3==1.34.69
botocore==1.34.69
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.4.0
urllib3==2.2.1
Werkzeug==3.0.1
xgboost==2.0.3
gunicorn==21.2.0